Split internal and FIRRTL packages
