+++ src/main/scala/Chisel/Bits.scala
@@ -1,8 +1,11 @@
 // See LICENSE for license details.
 
 package Chisel
-import Builder.pushOp
-import PrimOp._
+
+import internal._
+import internal.Builder.pushOp
+import firrtl._
+import firrtl.PrimOp._
 
 /** Element is a leaf data type: it cannot contain other Data objects. Example
   * uses are for representing primitive data types, like integers and bits.
