+++ src/main/scala/Chisel/Mem.scala
@@ -1,7 +1,10 @@
 // See LICENSE for license details.
 
 package Chisel
-import Builder.pushCommand
+
+import internal._
+import internal.Builder.pushCommand
+import firrtl._
 
 object Mem {
   @deprecated("Mem argument order should be size, t; this will be removed by the official release", "chisel3")
