+++ src/main/scala/Chisel/When.scala
@@ -1,7 +1,10 @@
 // See LICENSE for license details.
 
 package Chisel
-import Builder.pushCommand
+
+import internal._
+import internal.Builder.pushCommand
+import firrtl._
 
 object when {  // scalastyle:ignore object.name
   /** Create a `when` condition block, where whether a block of logic is
