+++ src/main/scala/Chisel/util/Conditional.scala
@@ -31,15 +31,15 @@ class SwitchContext[T <: Bits](cond: T) {
 object is {   // scalastyle:ignore object.name
   // Begin deprecation of non-type-parameterized is statements.
   def apply(v: Iterable[Bits])(block: => Unit) {
-    Builder.error("The 'is' keyword may not be used outside of a switch.")
+    require(false, "The 'is' keyword may not be used outside of a switch.")
   }
 
   def apply(v: Bits)(block: => Unit) {
-    Builder.error("The 'is' keyword may not be used outside of a switch.")
+    require(false, "The 'is' keyword may not be used outside of a switch.")
   }
 
   def apply(v: Bits, vr: Bits*)(block: => Unit) {
-    Builder.error("The 'is' keyword may not be used outside of a switch.")
+    require(false, "The 'is' keyword may not be used outside of a switch.")
   }
 }
